# Docs/kindohub-api-spec.yaml
openapi: 3.0.3
info:
  title: KindoHub API
  version: 1.0.0
  description: API REST para la gestion integral de una asociacion de padres y madres (APA/AMPA). Incluye gestion de familias, alumnos, cursos, anotaciones, usuarios, formas de pago, estados de asociado y sistema de auditoria con logs.
  contact:
    name: KindoHub Development Team
    email: dev@kindohub.com
  license:
    name: Proprietary
    url: https://kindohub.com/license
servers:
  - url: https://api.kindohub.com
    description: Servidor de produccion
  - url: https://staging-api.kindohub.com
    description: Servidor de staging
  - url: http://localhost:5000
    description: Desarrollo local
...
paths:
  /api/auth/login:
    post:
      tags:
        - Autenticacion
      summary: Iniciar sesion
      description: Autentica un usuario y retorna un token JWT
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenDto'
        '400':
          description: Datos invalidos
        '401':
          description: Credenciales incorrectas
        '500':
          description: Error interno del servidor
  /api/auth/logout:
    post:
      tags:
        - Autenticacion
      summary: Cerrar sesion
      description: Invalida el token de refresco del usuario
      operationId: logout
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout exitoso
        '401':
          description: No autenticado
        '500':
          description: Error interno del servidor
  /api/auth/refresh:
    post:
      tags:
        - Autenticacion
      summary: Refrescar token
      description: Obtiene un nuevo access token usando el refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefrescarTokenDto'
      responses:
        '200':
          description: Token refrescado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenDto'
        '401':
          description: Refresh token invalido o expirado
        '500':
          description: Error interno del servidor
...
  /api/alumnos/{id}:
    get:
      tags:
        - Alumnos
      summary: Obtener alumno por ID
      operationId: getStudentById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Alumno encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlumnoDto'
        '404':
          description: Alumno no encontrado
  /api/alumnos:
    get:
      tags:
        - Alumnos
      summary: Listar todos los alumnos
      operationId: getAllStudents
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de alumnos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AlumnoDto'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginDto:
      type: object
      required:
        - Username
        - Password
      properties:
        Username:
          type: string
          example: admin
        Password:
          type: string
          format: password
    TokenDto:
      type: object
      properties:
        Username:
          type: string
        AccessToken:
          type: string
        RefreshToken:
          type: string
        Roles:
          type: array
          items:
            type: string
    RefrescarTokenDto:
      type: object
      required:
        - RefreshToken
      properties:
        RefreshToken:
          type: string
...
    AlumnoDto:
      type: object
      properties:
        Id:
          type: integer
        Nombre:
          type: string
        FechaNacimiento:
          type: string
          format: date-time
        IdFamilia:
          type: integer
        IdCurso:
          type: integer
...
security:
  - BearerAuth: []